requirements.txt
streamlit>=1.27.0
ultralytics>=8.0.186
opencv-python>=4.8.0.76
Pillow>=9.5.0
numpy>=1.23.0
torch>=2.0.0
